state management, http error handling
