+++ src/app/auth/index.js
@@ -1,2 +1,2 @@
 export { default as authRoutes } from './auth-routes';
-export { default as authState } from './auth-state';
+export { default as authState } from './+store/auth-state';
