+++ src/app/courses/index.js
@@ -1,2 +1,2 @@
 export { default as courseRoutes } from './course-routes';
-export { default as courseState } from './course-state';
+export { default as courseState } from './+store/course-state';
