+++ src/app/shared/services/index.js
@@ -1,5 +1,3 @@
 export * from './httpClient';
-export * from './app-storage';
-export * from './app-logger';
 export * from './firebaseConfig';
 export * from './validators';
